Do not overwrite existing apt/php.list

+++ lamp.sh
@@ -25,8 +25,10 @@ export DEBIAN_FRONTEND=noninteractive
 install()
 {
     # Setup apt repository for php 8
-    wget -O /etc/apt/trusted.gpg.d/php.gpg https://packages.sury.org/php/apt.gpg
-    echo "deb https://packages.sury.org/php/ ${RELEASE%-security} main" > /etc/apt/sources.list.d/php.list
+    if [ ! -f /etc/apt/sources.list.d/php.list ] || [ ! -f /etc/apt/trusted.gpg.d/php.gpg ];then
+        wget -O /etc/apt/trusted.gpg.d/php.gpg https://packages.sury.org/php/apt.gpg
+        echo "deb https://packages.sury.org/php/ ${RELEASE%-security} main" > /etc/apt/sources.list.d/php.list
+    fi
     apt-get update
     $APTINSTALL apt-utils cron curl
     ls -l /var/lock || true
